% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{buildignore}
\alias{buildignore}
\title{Write .Rbuildignore}
\usage{
buildignore(projname = ".", add = FALSE,
  pat = c("^commit\\\\.command$", "\\\\.Rproj$", "^__.*",
  "^\\\\.DS_Store$"), verbose = TRUE)
}
\arguments{
\item{projname}{path to a package source, by default "." (current directory)}

\item{add}{logical; write new, or add to the existing file?}

\item{pat}{a character vector of regex patterns}

\item{verbose}{logical; print .Rbuildignore contents}
}
\description{
Write an .Rbuildignore file for a given package
}
