% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{desc_read}
\alias{desc_read}
\title{Read DESCRIPTION}
\usage{
desc_read(projname = ".", quiet = FALSE)
}
\arguments{
\item{projname}{path to a package source, by default "." (current directory)}

\item{quiet}{logical; if FALSE (default) results will be printed to console}
}
\description{
Read the DESCRIPTION file of a given package
}
\seealso{
\code{\link{desc_write}}, \code{\link{increment_version}}
}
