% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{desc_write}
\alias{desc_write}
\title{Write DESCRIPTION}
\usage{
desc_write(projname = ".", descv)
}
\arguments{
\item{projname}{path to a package source, by default "." (current directory)}

\item{descv}{a named character vector giving key-value pairs}
}
\description{
Write a DESCRIPTION file for a given package
}
\seealso{
\code{\link{desc_read}}, \code{\link{increment_version}}
}
