% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_xyz.R
\name{git_commit}
\alias{git_commit}
\title{Git commit}
\usage{
git_commit(projname = ".", msg, add = ".", push = FALSE)
}
\arguments{
\item{projname}{path to a folder with a git repository, by default "."
(current directory)}

\item{msg}{message to be included with the commit}

\item{add}{what files to include, either as a single character string, 
or a character vector, where each element is a directory, file or fileglob.}

\item{push}{logical; should the changes be pushed to a remote repository?}
}
\description{
Commit (and push) changes made to the git repository
}
\details{
This will add and commit, plus optionally push, all changes made in the given
repository. Push assumes a very basic setup and pushes upstream to the master
branch of 'origin'.
}
\seealso{
Other git functions: \code{\link{git_init}},
  \code{\link{git_push}}, \code{\link{git_remote}}
}
\concept{git functions}
