% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_xyz.R
\name{git_init}
\alias{git_init}
\title{Initialize git}
\usage{
git_init(projname = ".", add = ".", ignore = c(".DS_Store", ".RData",
  ".Rhistory", ".Rapp.history", ".Rproj.user"), commit = TRUE,
  msg = "first commit")
}
\arguments{
\item{projname}{path to a package source, by default "." (current directory)}

\item{add}{what files to include, by default "." (all)}

\item{ignore}{patterns to be written to .gitignore}

\item{commit}{logical; should changes be commited?}

\item{msg}{the message to be included with the commit}
}
\description{
Create an empty Git repository or reinitialize an existing one
}
\details{
An empty Git repository is created, files added, and a .gitignore file generated.
Optionally the changes are commited and stored with a message.
}
\seealso{
Other git functions: \code{\link{git_commit}},
  \code{\link{git_push}}, \code{\link{git_remote}}
}
\concept{git functions}
