% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_xyz.R
\name{git_remote}
\alias{git_remote}
\title{Add git remote}
\usage{
git_remote(projname = ".", repo)
}
\arguments{
\item{projname}{path to a folder with a git repository, by default "."
(current directory)}

\item{repo}{repository address in the format "username/repo[/subdir]"}
}
\description{
Add a remote at a given github repository
}
\details{
Connects a GitHub remote to the given git repository. This assumes a GitHub
repository and SSH is already set up.
}
\seealso{
Other git functions: \code{\link{git_commit}},
  \code{\link{git_init}}, \code{\link{git_push}}
}
\concept{git functions}
