% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{increment_version}
\alias{increment_version}
\title{Increment version}
\usage{
increment_version(projname = ".", comp = c("patch", "minor", "major"),
  dev = FALSE, quiet = FALSE)
}
\arguments{
\item{projname}{path to a package source, by default "." (current directory)}

\item{comp}{which component to increment. By default "patch"}

\item{dev}{logical; is it a development version?}

\item{quiet}{logical; if FALSE (default) results will be printed to console}
}
\description{
Increment version number in the DESCRIPTION file of a given package
}
\details{
Writes version numbers of the format "maj.min.patch.dev", but will also read
formats with hyphen seperated components, like "maj.min-patch". If \code{dev}
is \code{TRUE}, a version number of the form "maj.min.patch.9000" is used,
otherwise "maj.min.patch".
Other than version number, date, imports and R version number is updated as well.
}
\seealso{
\code{\link{desc_read}}, \code{\link{desc_write}}
}
