% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{params}
\alias{params}
\title{Function parameters}
\usage{
params(fun, to.clipboard = TRUE)
}
\arguments{
\item{fun}{name of function}

\item{to.clipboard}{logical; should results be copied to the clipboard?}
}
\description{
Return the parameters of a function, formatted for Roxygen2 use
}
\details{
The parameters are returned as "\code{@param <par>}" pairs, one per line, as
used by Roxygen2. The results can be either printed to console or copied to the
clipboard, in either case a character vector will be returned invisibly as well.
}
\examples{
params(params, FALSE)
params("params", FALSE)

}
\seealso{
\code{\link{roxdoc}}
}
