% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_xyz.R
\name{pkg_data}
\alias{pkg_data}
\title{Prepare datasets}
\usage{
pkg_data(projname = ".", add = FALSE)
}
\arguments{
\item{projname}{path to a package source, by default "." (current directory)}

\item{add}{logical; should datasets be added to those already existing in \code{data/}?}
}
\value{
\code{.rda} file(s) saved in the \code{data/} folder.
}
\description{
Save R objects as datasets
}
\details{
All R objects in \code{__data.R} are saved to the folder \code{data/} as files on the
pattern \code{d_<obj>.rda}, one for each object. If \code{add=FALSE}, then any existing
files of that pattern will first be removed, before new ones are saved.
}
\seealso{
\code{\link{pkg_objects}}
}
