% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_xyz.R
\name{pkg_install}
\alias{pkg_install}
\title{Install a package}
\usage{
pkg_install(projname = ".", bopt = "", iopt = "",
  ropt = c("--vanilla"), rm.src = TRUE)
}
\arguments{
\item{projname}{path to the project file}

\item{bopt}{\code{R CMD build} options}

\item{iopt}{\code{R CMD install} options}

\item{ropt}{\code{R} options}

\item{rm.src}{remove source file}
}
\description{
Install a package
}
