% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_xyz.R
\name{pkg_objects}
\alias{pkg_objects}
\title{Object sizes}
\usage{
pkg_objects(projname = ".")
}
\arguments{
\item{projname}{path to a package source, by default "." (current directory)}
}
\value{
A data.frame giving the size, name, class, source file name and folder for each object
}
\description{
Find the size of a package's loadable objects
}
\details{
The source code of a package is searched for function and dataset objects, found
in the \code{R/} and \code{dataset/} directories.
}
\seealso{
\code{\link{pkg_data}}
}
