% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_xyz.R
\name{pkg_pdf}
\alias{pkg_pdf}
\title{PDF manual}
\usage{
pkg_pdf(projname = ".", ropt = "--vanilla", popt = "--force")
}
\arguments{
\item{projname}{path to a package source, by default "." (current directory)}

\item{ropt}{a character vector, or space-delimited character string, of options
to be passed to \code{R}. By default \code{--vanilla} to run in a baseline R
session, reading no profile or environment files, and no saving of workspace}

\item{popt}{a character vector, or space-delimited character string, of options
to be passed to \code{\link{Rd2pdf}}. By default \code{--force} to overwrite existing
file}
}
\value{
A PDF file is saved to the root directory of the package
}
\description{
Render the PDF manual of all the package's documentation
}
\seealso{
\code{\link{Rd2pdf}}
}
