% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/0_doc_package.R
\docType{package}
\name{projector}
\alias{projector}
\alias{projector-package}
\title{projector: A short description of the package.}
\description{
A more detailed explanation/description of the package
}
\section{projector functions}{

\code{\link{pkg_check}} Short description
\code{\link{pkg_install}} Short description
}

