% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{r_manual}
\alias{r_manual}
\title{R manuals}
\usage{
r_manual(topic = c("extensions", "administration", "installation",
  "data", "I/O", "FAQ", "introduction", "internals",
  "language definition"))
}
\arguments{
\item{topic}{which help topic to browse. By default "extensions" ("R-exts.html")}
}
\description{
Browse offline R manuals
}
\details{
The AQUA browser will be used if AQUA is used as the GUI, otherwise
\code{getOption("browser")} is used.
}
