% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/etc.R
\name{roxdoc}
\alias{roxdoc}
\title{Documentation template}
\usage{
roxdoc(fun, to.clipboard = TRUE)
}
\arguments{
\item{fun}{function or name of function}

\item{to.clipboard}{logical; should results be copied to the clipboard?}
}
\value{
Either to the console, or clipboard, a function documentation template, as
appropriate for Roxygen2 use, is returned. Tags included are \code{@param}
(with argument names), \code{@details}, \code{@return}, \code{@seealso},
\code{@examples} and \code{@export}.
}
\description{
Create a function documentation template
}
\details{
The function arguments are returned as with \code{\link{params}}. 
In addition other commonly used Roxygen2 tags are included.
}
\examples{
roxdoc(roxdoc, FALSE)
roxdoc("roxdoc", FALSE)

}
\seealso{
\code{\link{params}}
}
