% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_xyz.R
\name{pkg_check}
\alias{pkg_check}
\title{Bild and run diagnostic checks on the package}
\usage{
pkg_check(projname = ".", bopt = c("--no-manual"),
  copt = c("--no-manual", "--timings"), ropt = c("--vanilla"),
  rm.src = TRUE)
}
\arguments{
\item{projname}{path to the project file}

\item{bopt}{\code{R CMD build} options}

\item{copt}{\code{R CMD check} options}

\item{ropt}{\code{R} options}

\item{rm.src}{remove source file}
}
\description{
Bild and run diagnostic checks on the package
}
